fix(*): fix algorithm to match instagrams more closely
The old version computed the probability for an item A based on the counts of all items. This made popular items score higher. Now the probability only depends on the item itelf which results in equal scoring, no matter the popularity.
The intuition for the current scoring is: out of this weeks events, how much % happened in the most recent bin.

Also the expectation and default values for new items are based on a sensible default.